+++ include/zephyr/drivers/gpio.h
@@ -863,67 +863,6 @@ __syscall int gpio_pin_interrupt_configure(const struct device *port,
 					   gpio_pin_t pin,
 					   gpio_flags_t flags);
 
-static inline int z_impl_gpio_pin_interrupt_configure(const struct device *port,
-						      gpio_pin_t pin,
-						      gpio_flags_t flags)
-{
-	const struct gpio_driver_api *api =
-		(const struct gpio_driver_api *)port->api;
-	__unused const struct gpio_driver_config *const cfg =
-		(const struct gpio_driver_config *)port->config;
-	const struct gpio_driver_data *const data =
-		(const struct gpio_driver_data *)port->data;
-	enum gpio_int_trig trig;
-	enum gpio_int_mode mode;
-
-	if (api->pin_interrupt_configure == NULL) {
-		return -ENOSYS;
-	}
-
-	__ASSERT((flags & (GPIO_INT_DISABLE | GPIO_INT_ENABLE))
-		 != (GPIO_INT_DISABLE | GPIO_INT_ENABLE),
-		 "Cannot both enable and disable interrupts");
-
-	__ASSERT((flags & (GPIO_INT_DISABLE | GPIO_INT_ENABLE)) != 0U,
-		 "Must either enable or disable interrupts");
-
-	__ASSERT(((flags & GPIO_INT_ENABLE) == 0) ||
-		 ((flags & GPIO_INT_EDGE) != 0) ||
-		 ((flags & (GPIO_INT_LOW_0 | GPIO_INT_HIGH_1)) !=
-		  (GPIO_INT_LOW_0 | GPIO_INT_HIGH_1)),
-		 "Only one of GPIO_INT_LOW_0, GPIO_INT_HIGH_1 can be "
-		 "enabled for a level interrupt.");
-
-	__ASSERT(((flags & GPIO_INT_ENABLE) == 0) ||
-#ifdef CONFIG_GPIO_ENABLE_DISABLE_INTERRUPT
-			 ((flags & (GPIO_INT_LOW_0 | GPIO_INT_HIGH_1)) != 0) ||
-			 (flags & GPIO_INT_ENABLE_DISABLE_ONLY) != 0,
-#else
-			 ((flags & (GPIO_INT_LOW_0 | GPIO_INT_HIGH_1)) != 0),
-#endif /* CONFIG_GPIO_ENABLE_DISABLE_INTERRUPT */
-		 "At least one of GPIO_INT_LOW_0, GPIO_INT_HIGH_1 has to be "
-		 "enabled.");
-
-	__ASSERT((cfg->port_pin_mask & (gpio_port_pins_t)BIT(pin)) != 0U,
-		 "Unsupported pin");
-
-	if (((flags & GPIO_INT_LEVELS_LOGICAL) != 0) &&
-	    ((data->invert & (gpio_port_pins_t)BIT(pin)) != 0)) {
-		/* Invert signal bits */
-		flags ^= (GPIO_INT_LOW_0 | GPIO_INT_HIGH_1);
-	}
-
-	trig = (enum gpio_int_trig)(flags & (GPIO_INT_LOW_0 | GPIO_INT_HIGH_1 | GPIO_INT_WAKEUP));
-#ifdef CONFIG_GPIO_ENABLE_DISABLE_INTERRUPT
-	mode = (enum gpio_int_mode)(flags & (GPIO_INT_EDGE | GPIO_INT_DISABLE | GPIO_INT_ENABLE |
-					     GPIO_INT_ENABLE_DISABLE_ONLY));
-#else
-	mode = (enum gpio_int_mode)(flags & (GPIO_INT_EDGE | GPIO_INT_DISABLE | GPIO_INT_ENABLE));
-#endif /* CONFIG_GPIO_ENABLE_DISABLE_INTERRUPT */
-
-	return api->pin_interrupt_configure(port, pin, mode, trig);
-}
-
 /**
  * @brief Configure pin interrupts from a @p gpio_dt_spec.
  *
@@ -961,58 +900,6 @@ static inline int gpio_pin_interrupt_configure_dt(const struct gpio_dt_spec *spe
 __syscall int gpio_pin_configure(const struct device *port,
 				 gpio_pin_t pin,
 				 gpio_flags_t flags);
-
-static inline int z_impl_gpio_pin_configure(const struct device *port,
-					    gpio_pin_t pin,
-					    gpio_flags_t flags)
-{
-	const struct gpio_driver_api *api =
-		(const struct gpio_driver_api *)port->api;
-	__unused const struct gpio_driver_config *const cfg =
-		(const struct gpio_driver_config *)port->config;
-	struct gpio_driver_data *data =
-		(struct gpio_driver_data *)port->data;
-
-	__ASSERT((flags & GPIO_INT_MASK) == 0,
-		 "Interrupt flags are not supported");
-
-	__ASSERT((flags & (GPIO_PULL_UP | GPIO_PULL_DOWN)) !=
-		 (GPIO_PULL_UP | GPIO_PULL_DOWN),
-		 "Pull Up and Pull Down should not be enabled simultaneously");
-
-	__ASSERT(!((flags & GPIO_INPUT) && !(flags & GPIO_OUTPUT) && (flags & GPIO_SINGLE_ENDED)),
-		 "Input cannot be enabled for 'Open Drain', 'Open Source' modes without Output");
-
-	__ASSERT_NO_MSG((flags & GPIO_SINGLE_ENDED) != 0 ||
-			(flags & GPIO_LINE_OPEN_DRAIN) == 0);
-
-	__ASSERT((flags & (GPIO_OUTPUT_INIT_LOW | GPIO_OUTPUT_INIT_HIGH)) == 0
-		 || (flags & GPIO_OUTPUT) != 0,
-		 "Output needs to be enabled to be initialized low or high");
-
-	__ASSERT((flags & (GPIO_OUTPUT_INIT_LOW | GPIO_OUTPUT_INIT_HIGH))
-		 != (GPIO_OUTPUT_INIT_LOW | GPIO_OUTPUT_INIT_HIGH),
-		 "Output cannot be initialized low and high");
-
-	if (((flags & GPIO_OUTPUT_INIT_LOGICAL) != 0)
-	    && ((flags & (GPIO_OUTPUT_INIT_LOW | GPIO_OUTPUT_INIT_HIGH)) != 0)
-	    && ((flags & GPIO_ACTIVE_LOW) != 0)) {
-		flags ^= GPIO_OUTPUT_INIT_LOW | GPIO_OUTPUT_INIT_HIGH;
-	}
-
-	flags &= ~GPIO_OUTPUT_INIT_LOGICAL;
-
-	__ASSERT((cfg->port_pin_mask & (gpio_port_pins_t)BIT(pin)) != 0U,
-		 "Unsupported pin");
-
-	if ((flags & GPIO_ACTIVE_LOW) != 0) {
-		data->invert |= (gpio_port_pins_t)BIT(pin);
-	} else {
-		data->invert &= ~(gpio_port_pins_t)BIT(pin);
-	}
-
-	return api->pin_configure(port, pin, flags);
-}
 
 /**
  * @brief Configure a single pin from a @p gpio_dt_spec and some extra flags.
@@ -1053,21 +940,6 @@ static inline int gpio_pin_configure_dt(const struct gpio_dt_spec *spec,
  */
 __syscall int gpio_port_get_direction(const struct device *port, gpio_port_pins_t map,
 				      gpio_port_pins_t *inputs, gpio_port_pins_t *outputs);
-
-#ifdef CONFIG_GPIO_GET_DIRECTION
-static inline int z_impl_gpio_port_get_direction(const struct device *port, gpio_port_pins_t map,
-						 gpio_port_pins_t *inputs,
-						 gpio_port_pins_t *outputs)
-{
-	const struct gpio_driver_api *api = (const struct gpio_driver_api *)port->api;
-
-	if (api->port_get_direction == NULL) {
-		return -ENOSYS;
-	}
-
-	return api->port_get_direction(port, map, inputs, outputs);
-}
-#endif /* CONFIG_GPIO_GET_DIRECTION */
 
 /**
  * @brief Check if @p pin is configured for input
@@ -1177,21 +1049,6 @@ static inline int gpio_pin_is_output_dt(const struct gpio_dt_spec *spec)
 __syscall int gpio_pin_get_config(const struct device *port, gpio_pin_t pin,
 				  gpio_flags_t *flags);
 
-#ifdef CONFIG_GPIO_GET_CONFIG
-static inline int z_impl_gpio_pin_get_config(const struct device *port,
-					     gpio_pin_t pin,
-					     gpio_flags_t *flags)
-{
-	const struct gpio_driver_api *api =
-		(const struct gpio_driver_api *)port->api;
-
-	if (api->pin_get_config == NULL)
-		return -ENOSYS;
-
-	return api->pin_get_config(port, pin, flags);
-}
-#endif
-
 /**
  * @brief Get a configuration of a single pin from a @p gpio_dt_spec.
  *
@@ -1229,15 +1086,6 @@ static inline int gpio_pin_get_config_dt(const struct gpio_dt_spec *spec,
  */
 __syscall int gpio_port_get_raw(const struct device *port,
 				gpio_port_value_t *value);
-
-static inline int z_impl_gpio_port_get_raw(const struct device *port,
-					   gpio_port_value_t *value)
-{
-	const struct gpio_driver_api *api =
-		(const struct gpio_driver_api *)port->api;
-
-	return api->port_get_raw(port, value);
-}
 
 /**
  * @brief Get logical level of all input pins in a port.
@@ -1293,16 +1141,6 @@ __syscall int gpio_port_set_masked_raw(const struct device *port,
 				       gpio_port_pins_t mask,
 				       gpio_port_value_t value);
 
-static inline int z_impl_gpio_port_set_masked_raw(const struct device *port,
-						  gpio_port_pins_t mask,
-						  gpio_port_value_t value)
-{
-	const struct gpio_driver_api *api =
-		(const struct gpio_driver_api *)port->api;
-
-	return api->port_set_masked_raw(port, mask, value);
-}
-
 /**
  * @brief Set logical level of output pins in a port.
  *
@@ -1348,15 +1186,6 @@ static inline int gpio_port_set_masked(const struct device *port,
 __syscall int gpio_port_set_bits_raw(const struct device *port,
 				     gpio_port_pins_t pins);
 
-static inline int z_impl_gpio_port_set_bits_raw(const struct device *port,
-						gpio_port_pins_t pins)
-{
-	const struct gpio_driver_api *api =
-		(const struct gpio_driver_api *)port->api;
-
-	return api->port_set_bits_raw(port, pins);
-}
-
 /**
  * @brief Set logical level of selected output pins to active.
  *
@@ -1386,15 +1215,6 @@ static inline int gpio_port_set_bits(const struct device *port,
 __syscall int gpio_port_clear_bits_raw(const struct device *port,
 				       gpio_port_pins_t pins);
 
-static inline int z_impl_gpio_port_clear_bits_raw(const struct device *port,
-						  gpio_port_pins_t pins)
-{
-	const struct gpio_driver_api *api =
-		(const struct gpio_driver_api *)port->api;
-
-	return api->port_clear_bits_raw(port, pins);
-}
-
 /**
  * @brief Set logical level of selected output pins to inactive.
  *
@@ -1423,15 +1243,6 @@ static inline int gpio_port_clear_bits(const struct device *port,
  */
 __syscall int gpio_port_toggle_bits(const struct device *port,
 				    gpio_port_pins_t pins);
-
-static inline int z_impl_gpio_port_toggle_bits(const struct device *port,
-					       gpio_port_pins_t pins)
-{
-	const struct gpio_driver_api *api =
-		(const struct gpio_driver_api *)port->api;
-
-	return api->port_toggle_bits(port, pins);
-}
 
 /**
  * @brief Set physical level of selected output pins.
@@ -1808,18 +1619,6 @@ static inline int gpio_remove_callback_dt(const struct gpio_dt_spec *spec,
  */
 __syscall int gpio_get_pending_int(const struct device *dev);
 
-static inline int z_impl_gpio_get_pending_int(const struct device *dev)
-{
-	const struct gpio_driver_api *api =
-		(const struct gpio_driver_api *)dev->api;
-
-	if (api->get_pending_int == NULL) {
-		return -ENOSYS;
-	}
-
-	return api->get_pending_int(dev);
-}
-
 /**
  * @}
  */
@@ -1828,6 +1627,7 @@ static inline int z_impl_gpio_get_pending_int(const struct device *dev)
 }
 #endif
 
+#include <zephyr/drivers/gpio/internal/gpio_impl.h>
 #include <zephyr/syscalls/gpio.h>
 
 #endif /* ZEPHYR_INCLUDE_DRIVERS_GPIO_H_ */
